Changed OnTimer's miliseconds argument type from unsigned long to __int64.
This makes current timestamp in miliseconds useful on 32-bit systems. Python: Changed OnTimer's argument from msecs of type int to seconds of type double.

+++ src/cconnpoll.h
@@ -54,7 +54,7 @@ public:
 	  */
 	virtual int Choose(nUtils::cTime &tmout)
 	{
-		return this->poll(tmout.MiliSec());
+		return this->poll((int)tmout.MiliSec());
 	};
 
 	/**
